Refactor experiments API routes into package
